# Reimplement lazy imports to remove use of codegen