… .pre-commit-config.yaml
… pyproject.toml
… src/globus_sdk/__init__.pyi
… src/globus_sdk/_generate_init.py
… tests/non-pytest/lazy-imports/test_no_repeat_names.py