This PR explores some new ideas about how to do our lazy imports.
The initial implementation was revised after a team discussion and review about strategy.

Benchmarking shows nearly equivalent performance to the codegen approach, as the AST traversals are quite fast.

----

The use of '_generate_init.py' makes the code unusual in a few ways which are not key to the lazy import semantics desired in the SDK.
Notably:
- we cannot use our typical code formatters on the init file
- the `__init__.py` cannot be edited by hand -- it must *always* be written via the codegen script
- the codegen script itself contains source-as-strings, which makes those pieces of functionality harder to alter and maintain

A new implementation based around AST-driven introspection provides equivalent semantics without any noticeable performance degradation.

## Initial Implementation

The idea is as follows:
- under the `if t.TYPE_CHECKING` branch, write relative imports as usual
- outside of that branch, provide a canonical `__all__` list to drive behaviors and identify "known names"
- in `__getattr__`, library-ize most of the lazy loading logic
- in that lazy-loading library, inspect module source and read the `if t.TYPE_CHECKING` branch
  - cache aggressively to avoid redoing AST work
  - scan it for `from .x import y` statements
  - redrive those parsed imports through `importlib`

## Revised Implementation

Following the above implementation, we took the same idea and tried using `.pyi` files to refine it further.
The ideas in the refinement are:
- we can load data out of an `__init__.pyi` and thereby avoid having to identify the `if t.TYPE_CHECKING` branch
- the internal lazy loading library can provide the implementations of `__dir__` and `__getattr__`, as well as the `__all__` tuple (for runtime use)
- now that we're locking in the implementation, more vigorous documentation effort is now appropriate

<!-- readthedocs-preview globus-sdk-python start -->
----
📚 Documentation preview 📚: https://globus-sdk-python--1100.org.readthedocs.build/en/1100/

<!-- readthedocs-preview globus-sdk-python end -->