@sirosen 2024-11-04
I've done a bit more testing and found that the original timing test I did was not correct.
I was timing how long it took to access an attribute of `globus_sdk`, but not clearing the module attribute for each run (so I was not _actually_ timing the codepath). Once corrected, the codegen approach properly shows itself to be faster, but note that this is down at nanoseconds scale.

---

Here's the simple benchmark script:
```python
import timeit

N = 10000

measurements = timeit.repeat(
    """
if hasattr(globus_sdk, "TransferClient"):
    del globus_sdk.TransferClient

globus_sdk.TransferClient""",
    setup="""
import globus_sdk
""",
    number=N,
)
print(f"benchmark: {min(measurements) / N * 1000_000_000}ns")
```
`benchmark.py`

Selecting the min of three runs for each implementation, I get:
- current (codegen): `benchmark: 1944.7158090770245ns`
- PR (AST): `benchmark: 11011.32296025753ns`

Both are exactly the same after the attribute is loaded (if you don't delete it).

In other words, the new approach is about 9000ns slower, or 0.009ms slower. I think that's an acceptable performance regression in order for us to have a more normal, and therefore maintainable, `__init__.py`.