@sirosen 2024-11-07
I've pushed the new implementation. With the new behavior, the benchmark above runs similarly:
```
benchmark: 10612.473799847066ns
```